rewrite/requirements.txt
Django>=4.2,<5.0
django-ninja>=1.0.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
sentry-sdk>=1.40.0
python-dateutil>=2.8.0
django-cors-headers>=4.3.0
requests>=2.31.0
Pillow>=10.0.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.5.0
